backend/services/rest-apis/requirements.txt
sqlalchemy
pymysql
mysql
mysqlclient
mysql-connector-python
boto3
bcrypt
pydantic
PyJWT